+++ cmd/cli/task_config/slot.go
@@ -1,4 +1,4 @@
-package config
+package task_config
 
 import (
 	"time"
